GEO-INFER-OPS/requirements.txt
pyyaml>=6.0
fastapi>=0.95.0
uvicorn>=0.21.0
prometheus-client>=0.16.0
structlog>=23.1.0
pytest>=7.3.1
docker>=6.0.1
kubernetes>=26.1.0

# Development dependencies
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pytest-cov>=4.1.0
pre-commit>=3.2.2 